requirements.txt
numpy==1.21.6
pandas==1.4.4
Pillow==9.4.0
requests==2.28.1
streamlit==1.20.0
ultralytics==8.0.57
